Implemented WinAppDriver to use the same test steps that WebDriver uses by creating a WindowsElement and overriding areas that were different. Added the ability to create executable paths in a page object instead or urls. Made it so the WinAppDriver.exe program is automatically run when needed. Assumes the path C:/Program Files (x86)/Windows Application Driver/WinAppDriver.exe though we could add a config option to allow it to be anywhere.
Currently the Windows programs run by the test close evn if the leaveBrowserOpen property is set to true. Need to implement leaveAppOpen property to spawn off the processes if necessary. Not sure how to do that right now.

Removed all references to WebDriverFactory and using it to start and stop webdrivers, and replaced it with a call to Driver as the interface for all drivers. Driver will spawn a different process for every differently named application, but only one WebDriver process for all web tests.