Updated save image code so that it returns the file created for operations that may require this output.
I also modified the save image code so it will verify the image has been saved to disk to avoid any instances where slow disk issues may result in a failure to find the image at a later step.

getImagePath has been modified so it can now accept a null string. Previously it would create a directory "null". Now it places the file in the root images folder as is expected behavior.