Add page and window to ensure there will not be any conflicts with page objects that include the word window or page.
Added custom message to the assertions.

Converted handling of windows vs page to a regex match rather than multiple to lower and equivalence.

Included new log messages for when black is used as a fallback (black chosen to indicate to user there is a resize issue due to lack of color information).

Logging will also issue a warning if a webpage is not being compared when the comparison images are of different sizes as the comparison may not be accurate.