Modified add new windows to return the number of new windows added for verification of an opened window
Modified prune closed windows to return the number of windows pruned for verification, and added logic to decrement the current window position in the list if the window closed was an earlier window. (race condition if a window opens while the previous closes and is pruned after the new window is fetched)

Go to new window logic added which will wait the default timeout for a new window to be added, and then set the current window to the end of the window handles list afterwards. Exception catching added for a default exception that the driver will throw with .getWindowHandles() if no windows exist (waiting on new window to load with previous closed)