Updated documentation
Changed the name of compareImages,
and modified the logic to handle the new multi-screenshot logic whereby it will only save results post-comparison, and can reuse the same temporary previous screenshot for multiple comparisons.

Logic updated to change the name of the output file including a rename based on pass/fail state, and process this pre-assertion.

Removed the expected/actual image writing (can be re-added if needed for debugging) as it could be misleading if the expected was not actually the expected state

Renamed some variables

Changed the logic to operate on true/false rather than ImageComparisonState in order to support some additional logic such as writing the result file with pass/fail before assertion.

Updated log messages