Add support for the chromium browser.
- This adds support for the open source chromium browser.
We install chromium into the GitHub workflow using chocolately (viz.
choco install chromium). By default chromedriver looks for chrome to be
installed in a standard location, which chrome is, so we override the
location for which to look for chome.exe, which is the chromium binary.

Add a basic sanity unit test that ensures the chromium browser can be
launched and made fullscreen.