Issue is a mix of things. For one, the getAttribute was failing to return a meaningful error for windows elements looking for color.
GetAttribute has been overridden in WindowsElement in order to return a special case if color is chosen. This will give us the value we're checking against in the error instead of "null" that the standard getAttribute returns.

For two, if notepad loses focus it no longer gets the correct color on hover.